refactor(hooks): move hooks directory to src/hooks
