+++ src/components/UISref.tsx
@@ -8,7 +8,7 @@ import * as _classNames from 'classnames';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 import { cloneElement, useCallback, useMemo } from 'react';
-import { useSref } from './hooks/useSref';
+import { useSref } from '../hooks/useSref';
 
 /** @hidden */
 let classNames = _classNames;
