+++ src/components/UISrefActive.tsx
@@ -8,7 +8,7 @@ import * as React from 'react';
 import { useState, useCallback, useContext, useMemo, cloneElement } from 'react';
 import * as _classNames from 'classnames';
 
-import { useCurrentStateAndParams, useRouter } from './hooks';
+import { useCurrentStateAndParams, useRouter } from '../hooks';
 
 /** @hidden */
 let classNames = _classNames;
