+++ src/components/__tests__/UIRouter.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import { mount } from 'enzyme';
 import { memoryLocationPlugin, UIRouter, UIRouterContext, UIRouterReact } from '../../index';
-import { muteConsoleErrors } from './util';
+import { muteConsoleErrors } from '../../__tests__/util';
 
 class Child extends React.Component<any, any> {
   static propTypes: React.ValidationMap<any> = {
