+++ src/components/index.ts
@@ -16,4 +16,3 @@ export * from './UIRouter';
 export * from './UIView';
 export * from './UISref';
 export * from './UISrefActive';
-export * from './hooks';
